split some tests
