+++ posthog/session_recordings/queries/test/listing_recordings/test_session_recording_list_from_query.py
@@ -1,4 +1,4 @@
-from datetime import datetime, UTC
+from datetime import datetime
 from typing import Literal
 from unittest.mock import ANY
 from uuid import uuid4
@@ -11,16 +11,15 @@ from parameterized import parameterized
 from posthog.clickhouse.client import sync_execute
 from posthog.clickhouse.log_entries import TRUNCATE_LOG_ENTRIES_TABLE_SQL
 from posthog.constants import AvailableFeature
-from posthog.hogql_queries.utils.query_date_range import QueryDateRange
-from posthog.models import Cohort, GroupTypeMapping, Person
+from posthog.models import GroupTypeMapping, Person
 from posthog.models.action import Action
 from posthog.models.group.util import create_group
 from posthog.models.team import Team
-from posthog.schema import RecordingsQuery, DateRange
+from posthog.schema import RecordingsQuery
+from posthog.session_recordings.queries.session_recording_list_from_query import SessionRecordingListFromQuery
 from posthog.session_recordings.queries.session_recording_list_from_query import (
     SessionRecordingQueryResult,
 )
-from posthog.session_recordings.queries.session_recording_list_from_query import SessionRecordingListFromQuery
 from posthog.session_recordings.queries.session_replay_events import ttl_days
 from posthog.session_recordings.queries.test.session_replay_sql import (
     produce_replay_summary,
@@ -37,74 +36,6 @@ from posthog.test.base import (
     flush_persons_and_events,
     snapshot_clickhouse_queries,
 )
-
-
-@freeze_time("2021-01-01T13:46:23")
-class TestSessionRecordingQueryDateRange(APIBaseTest):
-    def test_with_relative_dates(self) -> None:
-        query_date_range = QueryDateRange(
-            date_range=DateRange(date_from="-3d", date_to="-24h", explicitDate=True),
-            team=self.team,
-            interval=None,
-            now=datetime.now(UTC),
-        )
-
-        assert query_date_range.date_from() == datetime(2020, 12, 29, 0, 0, 0, 0, UTC)
-        assert query_date_range.date_to() == datetime(
-            year=2020, month=12, day=31, hour=13, minute=46, second=23, tzinfo=UTC
-        )
-
-    def test_with_string_dates(self) -> None:
-        query_date_range = QueryDateRange(
-            date_range=DateRange(date_from="2020-12-29", date_to="2021-01-01", explicitDate=True),
-            team=self.team,
-            interval=None,
-            now=datetime.now(UTC),
-        )
-
-        assert query_date_range.date_from() == datetime(2020, 12, 29, 0, 0, 0, 0, UTC)
-        assert query_date_range.date_to() == datetime(year=2021, month=1, day=1, hour=0, minute=0, second=0, tzinfo=UTC)
-
-    def test_with_string_date_times(self) -> None:
-        query_date_range = QueryDateRange(
-            date_range=DateRange(date_from="2020-12-29T12:23:45Z", date_to="2021-01-01T13:34:42Z", explicitDate=True),
-            team=self.team,
-            interval=None,
-            now=datetime.now(UTC),
-        )
-
-        assert query_date_range.date_from() == datetime(2020, 12, 29, 12, 23, 45, tzinfo=UTC)
-        assert query_date_range.date_to() == datetime(
-            year=2021, month=1, day=1, hour=13, minute=34, second=42, tzinfo=UTC
-        )
-
-    def test_with_no_date_from(self) -> None:
-        query_date_range = QueryDateRange(
-            date_range=DateRange(date_from=None, date_to="2021-01-01T13:34:42Z", explicitDate=True),
-            team=self.team,
-            interval=None,
-            now=datetime.now(UTC),
-        )
-
-        # defaults to start of 7 days ago
-        assert query_date_range.date_from() == datetime(2020, 12, 25, 0, 0, 0, 0, UTC)
-        assert query_date_range.date_to() == datetime(
-            year=2021, month=1, day=1, hour=13, minute=34, second=42, tzinfo=UTC
-        )
-
-    def test_with_no_date_to(self) -> None:
-        query_date_range = QueryDateRange(
-            date_range=DateRange(date_from="2021-01-01T11:34:42Z", date_to=None, explicitDate=True),
-            team=self.team,
-            interval=None,
-            now=datetime.now(UTC),
-        )
-
-        assert query_date_range.date_from() == datetime(2021, 1, 1, 11, 34, 42, tzinfo=UTC)
-        # defaults to now
-        assert query_date_range.date_to() == datetime(
-            year=2021, month=1, day=1, hour=13, minute=46, second=23, tzinfo=UTC
-        )
 
 
 @freeze_time("2021-01-01T13:46:23")
@@ -2207,395 +2138,6 @@ class TestSessionRecordingsListFromQuery(ClickhouseTestMixin, APIBaseTest):
             {"properties": [{"key": "email", "value": "gmail.com", "operator": "not_icontains", "type": "person"}]},
             [session_id_two],
         )
-
-    @snapshot_clickhouse_queries
-    @also_test_with_materialized_columns(person_properties=["$some_prop"])
-    def test_filter_with_cohort_properties(self) -> None:
-        with self.settings(USE_PRECALCULATED_CH_COHORT_PEOPLE=True):
-            with freeze_time("2021-08-21T20:00:00.000Z"):
-                user_one = "test_filter_with_cohort_properties-user"
-                user_two = "test_filter_with_cohort_properties-user2"
-                session_id_one = "session_not_in_cohort"
-                session_id_two = "session_in_cohort"
-
-                Person.objects.create(team=self.team, distinct_ids=[user_one], properties={"email": "bla"})
-                Person.objects.create(
-                    team=self.team,
-                    distinct_ids=[user_two],
-                    properties={"email": "bla2", "$some_prop": "some_val"},
-                )
-                cohort = Cohort.objects.create(
-                    team=self.team,
-                    name="cohort1",
-                    groups=[
-                        {
-                            "properties": [
-                                {
-                                    "key": "$some_prop",
-                                    "value": "some_val",
-                                    "type": "person",
-                                }
-                            ]
-                        }
-                    ],
-                )
-                cohort.calculate_people_ch(pending_version=0)
-
-                produce_replay_summary(
-                    distinct_id=user_one,
-                    session_id=session_id_one,
-                    first_timestamp=self.an_hour_ago,
-                    team_id=self.team.id,
-                )
-                # self.create_event(user_one, self.base_time, team=self.team)
-                produce_replay_summary(
-                    distinct_id=user_one,
-                    session_id=session_id_one,
-                    first_timestamp=self.an_hour_ago + relativedelta(seconds=30),
-                    team_id=self.team.id,
-                )
-                produce_replay_summary(
-                    distinct_id=user_two,
-                    session_id=session_id_two,
-                    first_timestamp=self.an_hour_ago,
-                    team_id=self.team.id,
-                )
-                # self.create_event(user_two, self.base_time, team=self.team)
-                produce_replay_summary(
-                    distinct_id=user_two,
-                    session_id=session_id_two,
-                    first_timestamp=self.an_hour_ago + relativedelta(seconds=30),
-                    team_id=self.team.id,
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": cohort.pk,
-                                "operator": "in",
-                                "type": "cohort",
-                            }
-                        ]
-                    },
-                    [session_id_two],
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": cohort.pk,
-                                "operator": "not_in",
-                                "type": "cohort",
-                            }
-                        ]
-                    },
-                    [session_id_one],
-                )
-
-    @snapshot_clickhouse_queries
-    @also_test_with_materialized_columns(person_properties=["$some_prop"])
-    def test_filter_with_static_and_dynamic_cohort_properties(self):
-        with self.settings(USE_PRECALCULATED_CH_COHORT_PEOPLE=True):
-            with freeze_time("2021-08-21T20:00:00.000Z"):
-                user_one = "test_filter_with_cohort_properties-user-in-static-cohort"
-                user_two = "test_filter_with_cohort_properties-user2-in-dynamic-cohort"
-                user_three = "test_filter_with_cohort_properties-user3-in-both-cohort"
-                user_four = "test_filter_with_cohort_properties-user4-not-in-any-cohort"
-
-                session_id_one = (
-                    f"in-static-cohort-test_filter_with_static_and_dynamic_cohort_properties-1-{str(uuid4())}"
-                )
-                session_id_two = (
-                    f"in-dynamic-cohort-test_filter_with_static_and_dynamic_cohort_properties-2-{str(uuid4())}"
-                )
-                session_id_three = (
-                    f"in-both-cohort-test_filter_with_static_and_dynamic_cohort_properties-3-{str(uuid4())}"
-                )
-                session_id_four = (
-                    f"not-in-any-cohort-test_filter_with_static_and_dynamic_cohort_properties-4-{str(uuid4())}"
-                )
-
-                Person.objects.create(team=self.team, distinct_ids=[user_one], properties={"email": "in@static.cohort"})
-                Person.objects.create(
-                    team=self.team,
-                    distinct_ids=[user_two],
-                    properties={"email": "in@dynamic.cohort", "$some_prop": "some_val"},
-                )
-                Person.objects.create(
-                    team=self.team,
-                    distinct_ids=[user_three],
-                    properties={"email": "in@both.cohorts", "$some_prop": "some_val"},
-                )
-
-                dynamic_cohort = Cohort.objects.create(
-                    team=self.team,
-                    name="cohort1",
-                    groups=[
-                        {
-                            "properties": [
-                                {
-                                    "key": "$some_prop",
-                                    "value": "some_val",
-                                    "type": "person",
-                                }
-                            ]
-                        }
-                    ],
-                )
-
-                static_cohort = Cohort.objects.create(team=self.team, name="a static cohort", groups=[], is_static=True)
-                static_cohort.insert_users_by_list([user_one, user_three])
-
-                dynamic_cohort.calculate_people_ch(pending_version=0)
-                static_cohort.calculate_people_ch(pending_version=0)
-
-                replay_summaries = [
-                    (user_one, session_id_one),
-                    (user_two, session_id_two),
-                    (user_three, session_id_three),
-                ]
-                for distinct_id, session_id in replay_summaries:
-                    produce_replay_summary(
-                        distinct_id=distinct_id,
-                        session_id=session_id,
-                        first_timestamp=self.an_hour_ago,
-                        team_id=self.team.id,
-                    )
-                    produce_replay_summary(
-                        distinct_id=distinct_id,
-                        session_id=session_id,
-                        first_timestamp=self.an_hour_ago + relativedelta(seconds=30),
-                        team_id=self.team.id,
-                    )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": static_cohort.pk,
-                                "operator": "in",
-                                "type": "cohort",
-                            },
-                        ]
-                    },
-                    [session_id_one, session_id_three],
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": static_cohort.pk,
-                                "operator": "not_in",
-                                "type": "cohort",
-                            },
-                        ]
-                    },
-                    [session_id_two],
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": dynamic_cohort.pk,
-                                "operator": "in",
-                                "type": "cohort",
-                            },
-                        ]
-                    },
-                    [session_id_two, session_id_three],
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": dynamic_cohort.pk,
-                                "operator": "not_in",
-                                "type": "cohort",
-                            },
-                        ]
-                    },
-                    [session_id_one],
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": dynamic_cohort.pk,
-                                "operator": "not_in",
-                                "type": "cohort",
-                            },
-                            {
-                                "key": "id",
-                                "value": static_cohort.pk,
-                                "operator": "not_in",
-                                "type": "cohort",
-                            },
-                        ]
-                    },
-                    [],
-                )
-
-                # and now with users not in any cohort
-
-                Person.objects.create(
-                    team=self.team, distinct_ids=[user_four], properties={"email": "not.in.any@cohorts.com"}
-                )
-                produce_replay_summary(
-                    distinct_id=user_four,
-                    session_id=session_id_four,
-                    team_id=self.team.id,
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": dynamic_cohort.pk,
-                                "operator": "not_in",
-                                "type": "cohort",
-                            },
-                            {
-                                "key": "id",
-                                "value": static_cohort.pk,
-                                "operator": "not_in",
-                                "type": "cohort",
-                            },
-                        ]
-                    },
-                    [session_id_four],
-                )
-
-    @snapshot_clickhouse_queries
-    @also_test_with_materialized_columns(person_properties=["$some_prop"])
-    def test_filter_with_events_and_cohorts(self):
-        with self.settings(USE_PRECALCULATED_CH_COHORT_PEOPLE=True):
-            with freeze_time("2021-08-21T20:00:00.000Z"):
-                user_one = "test_filter_with_events_and_cohorts-user"
-                user_two = "test_filter_with_events_and_cohorts-user2"
-                session_id_one = f"test_filter_with_events_and_cohorts-1-{str(uuid4())}"
-                session_id_two = f"test_filter_with_events_and_cohorts-2-{str(uuid4())}"
-
-                Person.objects.create(team=self.team, distinct_ids=[user_one], properties={"email": "bla"})
-                Person.objects.create(
-                    team=self.team,
-                    distinct_ids=[user_two],
-                    properties={"email": "bla2", "$some_prop": "some_val"},
-                )
-                cohort = Cohort.objects.create(
-                    team=self.team,
-                    name="cohort1",
-                    groups=[
-                        {
-                            "properties": [
-                                {
-                                    "key": "$some_prop",
-                                    "value": "some_val",
-                                    "type": "person",
-                                }
-                            ]
-                        }
-                    ],
-                )
-                cohort.calculate_people_ch(pending_version=0)
-
-                produce_replay_summary(
-                    distinct_id=user_one,
-                    session_id=session_id_one,
-                    first_timestamp=self.an_hour_ago,
-                    team_id=self.team.id,
-                )
-                self.create_event(
-                    user_one,
-                    self.an_hour_ago,
-                    team=self.team,
-                    event_name="custom_event",
-                    properties={"$session_id": session_id_one},
-                )
-                produce_replay_summary(
-                    distinct_id=user_one,
-                    session_id=session_id_one,
-                    first_timestamp=self.an_hour_ago + relativedelta(seconds=30),
-                    team_id=self.team.id,
-                )
-                produce_replay_summary(
-                    distinct_id=user_two,
-                    session_id=session_id_two,
-                    first_timestamp=self.an_hour_ago,
-                    team_id=self.team.id,
-                )
-                self.create_event(
-                    user_two,
-                    self.an_hour_ago,
-                    team=self.team,
-                    event_name="custom_event",
-                    properties={"$session_id": session_id_two},
-                )
-                produce_replay_summary(
-                    distinct_id=user_two,
-                    session_id=session_id_two,
-                    first_timestamp=self.an_hour_ago + relativedelta(seconds=30),
-                    team_id=self.team.id,
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        # has to be in the cohort and pageview has to be in the events
-                        # test data has one user in the cohort but no pageviews
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": cohort.pk,
-                                "operator": "in",
-                                "type": "cohort",
-                            }
-                        ],
-                        "events": [
-                            {
-                                "id": "$pageview",
-                                "type": "events",
-                                "order": 0,
-                                "name": "$pageview",
-                            }
-                        ],
-                    },
-                    [],
-                )
-
-                self._assert_query_matches_session_ids(
-                    {
-                        "properties": [
-                            {
-                                "key": "id",
-                                "value": cohort.pk,
-                                "operator": "in",
-                                "type": "cohort",
-                            }
-                        ],
-                        "events": [
-                            {
-                                "id": "custom_event",
-                                "type": "events",
-                                "order": 0,
-                                "name": "custom_event",
-                            }
-                        ],
-                    },
-                    [session_id_two],
-                )
 
     @snapshot_clickhouse_queries
     @also_test_with_materialized_columns(["$current_url"])
